task2/requirements.txt
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.3
PyYAML==6.0.2